refactor(file naming): rename salt-rock.js to saltRock.js
