+++ src/data/ores/index.js
@@ -4,4 +4,4 @@ export { goldOre } from './goldOre'
 export { ironOre } from './ironOre'
 export { silverOre } from './silverOre'
 export { stone } from './stone'
-export { saltRock } from './salt-rock'
+export { saltRock } from './saltRock'
