+++ src/data/ores/saltRock.js
